Extended file functionality
Almost all files work on the main code except file t1-1c, this is due to an error where 2 non blocking grass elements are placed on the same tile.

I have added a rabbit and burrow class. Grass now implements non-blocking so that rabbit can stand on the same tile as grass.